chore: rename lib subdirs
