+++ .gitignore
@@ -12,7 +12,7 @@ yarn-error.log
 lerna-debug.log
 
 # compile source
-/lib
+lib
 
 # test artifacts
 *xunit.xml
