+++ src/shared/community/connect/CommunityTemplatesResource.ts
