+++ src/shared/community/defs/CommunityInfo.ts
