+++ src/shared/community/defs/CommunityTemplates.ts
