+++ src/shared/community/service/CommunitiesServices.ts
@@ -4,7 +4,7 @@
  * Licensed under the BSD 3-Clause license.
  * For full license text, see LICENSE.txt file in the repo root or https://opensource.org/licenses/BSD-3-Clause
  */
-import { Org } from '@salesforce/core/lib/org';
+import { Org } from '@salesforce/core';
 import { QueryResult } from 'jsforce';
 import { CommunityInfo } from '../defs/CommunityInfo';
