+++ src/shared/connect/services/ConnectResource.ts
